Only intercept hooks when requested
